implement final task (MeetingTime & Payment) 🫡

+++ Data_Access_Layer/DTOs/Payment_DTOs/PaymentDto.cs
@@ -8,7 +8,7 @@ namespace Data_Access.DTOs.Payment_DTOs
         public int? PaymentId { get; set; }
         public int StudentsID { get; set; }
         public int GroupId { get; set; }
-        public int PaymentAmount { get; set; }
+        public decimal PaymentAmount { get; set; }
         public int SubjectGradeLevelId { get; set; }
         public DateTime? PaymentDate { get; set; }
